refact: 해몽에서 엔티티 dto 변환 로직 분리
