+++ Backend/src/main/java/dsko/hier/fortune/dream/domain/DreamAnalysisRepository.java
@@ -1,4 +1,4 @@
-package dsko.hier.fortune.dreamInterpretation.domain;
+package dsko.hier.fortune.dream.domain;
 
 public interface DreamAnalysisRepository {
     DreamAnalysis save(DreamAnalysis dreamAnalysis);
