+++ Backend/src/main/java/dsko/hier/fortune/dream/infra/DreamAnalysisJpaRepository.java
@@ -1,6 +1,6 @@
-package dsko.hier.fortune.dreamInterpretation.infra;
+package dsko.hier.fortune.dream.infra;
 
-import dsko.hier.fortune.dreamInterpretation.domain.DreamAnalysis;
+import dsko.hier.fortune.dream.domain.DreamAnalysis;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
