+++ Backend/src/main/java/dsko/hier/fortune/dream/infra/DreamAnalysisRepositoryImpl.java
@@ -1,7 +1,7 @@
-package dsko.hier.fortune.dreamInterpretation.infra;
+package dsko.hier.fortune.dream.infra;
 
-import dsko.hier.fortune.dreamInterpretation.domain.DreamAnalysis;
-import dsko.hier.fortune.dreamInterpretation.domain.DreamAnalysisRepository;
+import dsko.hier.fortune.dream.domain.DreamAnalysis;
+import dsko.hier.fortune.dream.domain.DreamAnalysisRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 public class DreamAnalysisRepositoryImpl implements DreamAnalysisRepository {
 
     private final DreamAnalysisJpaRepository repository;
-    
+
     @Override
     public DreamAnalysis save(DreamAnalysis dreamAnalysis) {
         return repository.save(dreamAnalysis);
